Change int to float.

+++ log2grasp.py
@@ -23,24 +23,24 @@ queues = {}
 for line in lines :
 	line = line.strip()
 	inst, args = line.split(' ', 1)
-	
+
 	if inst == 'task' :
 		id, priority, name = args.split(' ', 2)
-		
+
 		task = {}
 		task['no'] = str(len(tasks) + 1)
 		task['priority'] = int(priority)
 		task['name'] = task['no'] + ": " + name.strip()
 		task['created'] = True
-		
+
 		tasks[id] = task
-		
+
 	elif inst == 'switch' :
 		out_task, in_task, tick, tick_reload, out_minitick, in_minitick = args.split(' ')
-		
-		out_time = (int(tick) + (int(tick_reload) - int(out_minitick)) / int(tick_reload)) / 100 * 1000;
-		in_time  = (int(tick) + (int(tick_reload) - int(in_minitick))  / int(tick_reload)) / 100 * 1000;
-		
+
+		out_time = (float(tick) + (float(tick_reload) - float(out_minitick)) / float(tick_reload)) / 100 * 1000;
+		in_time  = (float(tick) + (float(tick_reload) - float(in_minitick))  / float(tick_reload)) / 100 * 1000;
+
 		event = {}
 		event['type'] = 'task out'
 		event['task'] = out_task
@@ -121,7 +121,7 @@ for line in lines :
 					tasks[task_id] = task
 
 				events.append(event);
-		
+
 		elif act == 'block' :
 			time, task_id, id = args.split(' ')
 			if id in all_queues and all_queues[id]['type'] == 'binary semaphore':
@@ -231,7 +231,7 @@ for event in events :
 	elif event['type'] == 'interrupt out' :
 		grasp.write('plot %f jobCompleted job%s.1\n' % (event['time'], event['task']))
 		grasp.write('plot %f jobResumed job%s.1\n' % (event['time'], event['prev']))
-		
+
 # Clean up unended operations
 
 for id in mutexes :
